Given a string , decode the password by following the steps mentioned.
Problem statement : Aman, who is working at a software company forgot the password of his Linkedin id.But he knows the ASCII values of his password in reverse order. Help aman to find the password.
To decode the password, first reverse the string of digits, then successively pick valid values from the string and convert them to their ASCII equivalents. Some of the values will have two digits, and others three. Use the ranges of valid values when decoding the string of digits.

Some of the ASCII values are given with their characters:

 The ASCII value of A to Z is 65 to 90.
 The ASCII value of a to z is 97 to 122.
 The ASCII value of space characters is 32.
Note: The password only has alphabets and blank spaces.